refactor(test): extract adminAuthSetup helper for campaign tests. Fixes #619
