+++ apps/web/app/api/admin/campaigns/[id]/preview/route.test.ts
@@ -21,9 +21,7 @@ vi.mock("@/lib/email/templates/announcement", () => ({
   buildAnnouncementHtml: vi.fn(() => "<html><body>Preview</body></html>"),
 }));
 
-import { readSessionCookie } from "@/lib/auth/github";
-import { isAdminHandle } from "@/lib/auth/admin";
-import { rateLimit } from "@/lib/cache/redis";
+import { adminAuthBeforeEach, readSessionCookie, isAdminHandle, rateLimit } from "@/lib/test-helpers/admin-auth";
 import { dbGetCampaign } from "@/lib/db/campaigns";
 import { buildAnnouncementHtml } from "@/lib/email/templates/announcement";
 import { GET } from "./route";
@@ -31,12 +29,7 @@ import { GET } from "./route";
 const mockParams = Promise.resolve({ id: "c-1" });
 
 beforeEach(() => {
-  vi.clearAllMocks();
-  process.env.NEXTAUTH_SECRET = "test-secret";
-  vi.mocked(readSessionCookie).mockReturnValue({
-    token: "t", login: "admin", name: "Admin", avatar_url: "",
-  });
-  vi.mocked(isAdminHandle).mockReturnValue(true);
+  adminAuthBeforeEach();
 });
 
 function makeRequest(): NextRequest {
