+++ apps/web/app/api/admin/campaigns/[id]/route.test.ts
@@ -20,21 +20,14 @@ vi.mock("@/lib/db/campaigns", () => ({
   dbDeleteCampaign: vi.fn(),
 }));
 
-import { readSessionCookie } from "@/lib/auth/github";
-import { isAdminHandle } from "@/lib/auth/admin";
-import { rateLimit } from "@/lib/cache/redis";
+import { adminAuthBeforeEach, readSessionCookie, isAdminHandle, rateLimit } from "@/lib/test-helpers/admin-auth";
 import { dbGetCampaign, dbUpdateCampaign, dbDeleteCampaign } from "@/lib/db/campaigns";
 import { GET, PATCH, DELETE } from "./route";
 
 const mockParams = Promise.resolve({ id: "c-1" });
 
 beforeEach(() => {
-  vi.clearAllMocks();
-  process.env.NEXTAUTH_SECRET = "test-secret";
-  vi.mocked(readSessionCookie).mockReturnValue({
-    token: "t", login: "admin", name: "Admin", avatar_url: "",
-  });
-  vi.mocked(isAdminHandle).mockReturnValue(true);
+  adminAuthBeforeEach();
 });
 
 function makeRequest(method = "GET", body?: unknown): NextRequest {
