+++ apps/web/app/api/admin/campaigns/[id]/test/route.test.ts
@@ -32,9 +32,7 @@ vi.mock("@/lib/email/templates/announcement", () => ({
   buildAnnouncementText: vi.fn().mockReturnValue("test"),
 }));
 
-import { readSessionCookie } from "@/lib/auth/github";
-import { isAdminHandle } from "@/lib/auth/admin";
-import { rateLimit } from "@/lib/cache/redis";
+import { adminAuthBeforeEach, readSessionCookie, isAdminHandle, rateLimit } from "@/lib/test-helpers/admin-auth";
 import { requireSession } from "@/lib/auth/require-session";
 import { dbGetCampaign } from "@/lib/db/campaigns";
 import { getResend } from "@/lib/email/resend";
@@ -64,12 +62,7 @@ const CAMPAIGN = {
 };
 
 beforeEach(() => {
-  vi.clearAllMocks();
-  process.env.NEXTAUTH_SECRET = "test-secret";
-  vi.mocked(readSessionCookie).mockReturnValue({
-    token: "t", login: "admin", name: "Admin", avatar_url: "",
-  });
-  vi.mocked(isAdminHandle).mockReturnValue(true);
+  adminAuthBeforeEach();
   vi.mocked(requireSession).mockReturnValue({
     session: { token: "t", login: "admin", name: "Admin", avatar_url: "" },
   });
