+++ apps/web/app/api/admin/campaigns/route.test.ts
@@ -19,19 +19,12 @@ vi.mock("@/lib/db/campaigns", () => ({
   dbCreateCampaign: vi.fn(),
 }));
 
-import { readSessionCookie } from "@/lib/auth/github";
-import { isAdminHandle } from "@/lib/auth/admin";
-import { rateLimit } from "@/lib/cache/redis";
+import { adminAuthBeforeEach, readSessionCookie, isAdminHandle, rateLimit } from "@/lib/test-helpers/admin-auth";
 import { dbGetCampaigns, dbCreateCampaign } from "@/lib/db/campaigns";
 import { GET, POST } from "./route";
 
 beforeEach(() => {
-  vi.clearAllMocks();
-  process.env.NEXTAUTH_SECRET = "test-secret";
-  vi.mocked(readSessionCookie).mockReturnValue({
-    token: "t", login: "admin", name: "Admin", avatar_url: "",
-  });
-  vi.mocked(isAdminHandle).mockReturnValue(true);
+  adminAuthBeforeEach();
 });
 
 function makeRequest(method = "GET", body?: unknown): NextRequest {
